pyproject.toml
[project]
name = "joblens"
version = "0.1.0"
description = "Natural language analytics over job-market data using AWS Bedrock"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.42.42",
    "fastapi>=0.128.1",
    "langchain==1.2.8",
    "langchain-aws>=1.2.2",
    "langchain-community==0.4.1",
    "langchain-core==1.2.8",
    "langgraph==1.0.7",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.40.0",
]
